added microservices auth and fixed frontend

+++ shipments/src/main/java/ru/rsoi/shipments/web/ShipmentController.java
@@ -41,49 +41,70 @@ public class ShipmentController {
     }
 
     @GetMapping("/{id}")
-    public ShipmentInfo ShipmentById(@PathVariable Integer id) {
-        return shipmentService.getById(id);
+    public ResponseEntity<ShipmentInfo> ShipmentById(@PathVariable Integer id,@RequestHeader("token") String jwt) {
+        if (shipmentService.parseJWT(jwt)){
+        return ResponseEntity.ok(shipmentService.getById(id));}
+        else return ResponseEntity.status(401).body(null);
     }
 
     @GetMapping
-    public Page<ShipmentInfo> findAllShipments(@RequestParam(value = "page", defaultValue = "0") Integer page,@RequestParam(value="size",defaultValue = "2") Integer size) {
-
+    public ResponseEntity<Page<ShipmentInfo>> findAllShipments(@RequestParam(value = "page", defaultValue = "0") Integer page,@RequestParam(value="size",defaultValue = "2") Integer size,@RequestHeader("token") String jwt) {
+        if (shipmentService.parseJWT(jwt)){
         Pageable request = PageRequest.of(page,size);
-        return shipmentService.getAll(request);
+        return ResponseEntity.ok(shipmentService.getAll(request));}
+        else return ResponseEntity.status(401).body(null);
     }
 
     @PostMapping("/create")
-    public void createShipment(@RequestBody ShipmentInfo shipment) {
+    public ResponseEntity<Void> createShipment(@RequestBody ShipmentInfo shipment,@RequestHeader("token") String jwt) {
+        if (shipmentService.parseJWT(jwt)){
         shipmentService.createShipment(shipment);
+            return ResponseEntity.ok().build();}
+        else return ResponseEntity.status(401).build();
     }
 
     @PostMapping("/createAgr")
-    public void createShipmentAgr(@RequestBody JSONArray shipment) throws ParseException
+    public ResponseEntity<Void> createShipmentAgr(@RequestBody JSONArray shipment,@RequestHeader("token") String jwt) throws ParseException
     {
+        if (shipmentService.parseJWT(jwt)){
        // List<ShipmentInfo> new_shipment = shipmentService.getModelFromHashMap(shipment);
         shipmentService.createShipments(shipment);
+            return ResponseEntity.ok().build();}
+        else return ResponseEntity.status(401).build();
     }
 
     @PostMapping("/edit")
-    public void editShipment(@RequestBody ShipmentInfo shipment) {
-        shipmentService.editShipment(shipment);
+    public ResponseEntity<Void> editShipment(@RequestBody ShipmentInfo shipment,@RequestHeader("token") String jwt) {
+        if (shipmentService.parseJWT(jwt)){
+            shipmentService.editShipment(shipment);
+            return ResponseEntity.ok().build();}
+        else return ResponseEntity.status(401).build();
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteShipment(@PathVariable long id) {
+    public ResponseEntity<Void> deleteShipment(@PathVariable long id,@RequestHeader("token") String jwt) {
+        if (shipmentService.parseJWT(jwt)){
         shipmentService.delete(id);
+        return ResponseEntity.ok().build();}
+        else return ResponseEntity.status(401).build();
     }
 
     @GetMapping("/deliveries/{del_id}")
-    public List<ShipmentInfo> findAllByDeliveryId(@PathVariable Integer del_id)
+    public ResponseEntity<List<ShipmentInfo>> findAllByDeliveryId(@PathVariable Integer del_id,@RequestHeader("token") String jwt)
     {
-        return shipmentService.findAllByDeliveryId(del_id);
+        if (shipmentService.parseJWT(jwt)){
+        return ResponseEntity.ok(shipmentService.findAllByDeliveryId(del_id));}
+        else return ResponseEntity.status(401).body(null);
     }
 
     @DeleteMapping("/deleteAll/{id}")
-    public void deleteAllByDeliveryId(@PathVariable Integer id)
+    public ResponseEntity<Void> deleteAllByDeliveryId(@PathVariable Integer id,@RequestHeader("token") String jwt)
     {
+        if (shipmentService.parseJWT(jwt)){
         shipmentService.deleteAllByDeliveryId(id);
+        return ResponseEntity.ok().build();
+        }
+        else return ResponseEntity.status(401).build();
     }
